refactoring: move tests from ast to separate file
